requirements/dev.requirements.txt
-r requirements.txt
django-debug-toolbar==1.2.2
selenium==3.0.2
django-nose==1.4.2
coverage==3.6

model_mommy==1.2.6

splinter==0.7.5
django-dynamic-fixture==1.8.0

flake8==2.4.1
django-jenkins>=0.16.3
pep8==1.6.2
pyflakes==0.8.1

bumpversion==0.5.3

ipython==3.1.0
pyreadline==2.0

pytest==3.0.5
pytest-django==3.1.2
pytest-xdist==1.13.1
pytest-splinter==1.7.8
pytest-instafail==0.3.0
pytest-localserver==0.3.6
pdbpp

mock==1.0.1

django-webtest==1.8.0
beautifulsoup4==4.4.1
waitress==0.8.10
webtest==2.0.23
WebOb==1.6.3
